+++ src/main/java/at/grueneis/spengergasse/registry5/Parent.java
@@ -1,0 +1,7 @@
+package at.grueneis.spengergasse.registry5;
+
+/**
+ * Created by Daniel on 22/01/14.
+ */
+public class Parent {
+}
